fix some build problems:   add includes for qt 5.15.12   add actual log_level for boost 1.83.0   add static_cast to enum class Type

+++ simulator/runtime/calc/operation/calc_binary.h
@@ -44,7 +44,7 @@ friend class rdo::Factory<RDOCalcBinary<F, CalcType>>;
 public:
     enum class Type
     {
-        calc_type = CalcType
+        calc_type = static_cast<int>(CalcType)
     };
     typedef F caller_type;
